backend/packages/fiab-plugin-anemoi/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"

requires = [ "setuptools>=80", "setuptools-scm[toml]>=6.2" ]

[project]
name = "fiab-plugin-anemoi"
description = "Weather Forecast in a Box, Anemoi plugin"
readme = "README.md"
license = "Apache-2.0"
license-files = [ "LICENSE" ]
authors = [
  { name = "European Centre for Medium-Range Weather Forecasts (ECMWF)", email = "software.support@ecmwf.int" },
]
requires-python = ">=3.11"

dynamic = [ "version" ]

dependencies = [
    "fiab-plugin-ecmwf",
    "earthkit-workflows-anemoi>=0.4.0",
]

optional-dependencies.runtime = [

]

[dependency-groups]
dev = [
  "ty==0.0.2",
]

[tool.setuptools_scm]
version_file = "src/fiab_plugin_anemoi/_version.py"
write_to_template = '''# Do not change! Do not track in version control!
__version__ = "{version}"
'''
local_scheme = "no-local-version"
root = ".."
fallback_version = "0.0.0"
tag_regex = '^[vd](?P<version>\d+(?:\.\d+){0,2}[^\+]*)(?:\+.*)?$'

[tool.ruff]
line-length = 140

lint.ignore = [ "E731" ]